QF-Protocol/qf_protocol/requirements.txt
torch==2.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pycryptodome==3.19.0
cryptography==41.0.7
scipy==1.11.4
numpy==1.24.3
pydantic==2.5.0